It's A Chess Puzzle
This script was developed to function as a generalization of a chess puzzle found here:
https://github.com/mwerner640/chess_puzzle

My script attempts to address the case of any nxn square checkerboard, rather than just an 8x8 board.

My solution, though not as elegant as the solution linked above, does allow for a proper treatment for finding the conditions in which this type of puzzle has a solution, which I attempt to show visually through pygame.